op-supervisor: Move resuming databases into ChainsDB so it can later update the chain heads too.
